requirements.txt
# ──────────────────────────────────────────────────────────────
# DCGFM – OFA + GraphCLIP pipeline requirements
# GPU: NVIDIA RTX 5090 (sm_120 / Blackwell)
# CUDA: 12.8   Python: 3.10
# ──────────────────────────────────────────────────────────────

# ── PyTorch (nightly cu128 required for RTX 5090 / sm_120) ──
# Install separately BEFORE this file:
#   pip install --pre torch torchvision torchaudio \
#       --index-url https://download.pytorch.org/whl/nightly/cu128

# ── PyTorch Geometric extensions (built from source) ──
# These CANNOT be installed via pip index for nightly torch.
# Build script is provided in docker/build_pyg.py (runs automatically in Dockerfile).
# torch-scatter
# torch-sparse
# torch-cluster
# torch-spline-conv
torch-geometric==2.7.0

# ── Graph / ML core ──
ogb==1.3.6
dgl==2.5.0+cu124                    # install via: pip install dgl -f https://data.dgl.ai/wheels/torch-2.6/cu124/repo.html
scikit-learn==1.7.2
scipy==1.15.3
numpy==1.26.4                       # keep <2 for compatibility with compiled extensions
networkx==3.4.2
pandas==2.3.3

# ── Transformers / LLM stack ──
transformers==5.1.0
tokenizers==0.22.2
sentence-transformers==5.2.3
sentencepiece==0.2.1
peft==0.18.1
accelerate==1.12.0
bitsandbytes==0.49.1
safetensors==0.7.0
huggingface_hub==1.4.1
datasets==4.5.0

# ── Training framework ──
lightning==2.6.1
pytorch-lightning==2.6.1
torchmetrics==1.8.2
deepspeed==0.18.7

# ── Chemistry (OFA mol tasks) ──
rdkit==2025.9.6

# ── Logging & experiment tracking ──
wandb==0.25.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
test_tube==0.7.5

# ── Utilities ──
tqdm==4.67.3
pyyaml==6.0.3
hjson==3.1.0
ninja==1.13.0
psutil==7.2.2
gpustat==1.1.1
py-cpuinfo==9.0.0
pympler==1.1
nvidia-ml-py==13.590.48
matplotlib==3.10.8
pillow==12.1.0
requests==2.32.5
filelock==3.20.3
packaging==26.0
